Split unit tests and cli tests
