+++ test/cli/commands/util.ts
