+++ test/unit/util/config.test.ts
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 import { it, describe } from 'mocha';
 
-import ConnectionConfig from '../../src/domain/ConnectionConfig';
-import { validate, getConnectionId, resolveConnectionsFromEnv, isCLI } from '../../src/config';
+import ConnectionConfig from '../../../src/domain/ConnectionConfig';
+import { validate, getConnectionId, resolveConnectionsFromEnv, isCLI } from '../../../src/config';
 
 describe('CONFIG:', () => {
   describe('isCLI', () => {
